Project_server/requirements.txt
bidict==0.22.1
click==8.1.8
dnspython==2.3.0
eventlet==0.33.3
Flask==2.2.5
Flask-SocketIO==5.1.1
greenlet==3.1.1
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.6
MarkupSafe==2.1.5
pkg_resources==0.0.0
pyserial==3.5
python-engineio==4.3.4
python-socketio==5.5.2
six==1.17.0
typing_extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0
